refactor(fonts): address PR #190 review — fix timestamps, extract shared helpers, improve robustness
- Use CURRENT_TIMESTAMP for font created_at and published_at instead of
  Rust-side Utc::now(), consistent with the rest of the codebase
- Extract test-mode workspace bootstrap from font_handlers.rs and
  upload.rs into workspace::ensure_test_mode_workspace(), eliminating
  ~100 lines of duplicated code
- Extract read_font_row() helper to deduplicate the 13-column FontItem
  row mapping used in list_fonts and get_font
- Log DB errors in update_font_error instead of silently discarding them
- Save uploaded font with original file extension (e.g. original.ttf)
  for easier filesystem inspection
